Broadcast in inception finished
On switch connection, controller identify all ports of the switch.
Then a flow will be set on the switch.
On source learning, controller will set up another flow on the switch
to override the first flow when the source of the broadcast message
is from hosts connected by the switch